[std] completely rewrite module 'ascii'
A completely new API is provided, with useful documentation and proper
typing.  Currently, functions for ASCII strings are not provided -- they
require more consideration in terms of interface and judging what would
be idiomatic.

This breaks every current use of the 'ascii' module.  These usages will
be fixed as those modules are rewritten.