[std/ascii] port 'std' to the new API
These are fundamentally incomplete ports, because they only update calls
to ASCII functions; the new API exposes an ASCII character type which is
meant to be used for all ASCII data, and the ported uses in the rest of
'std' do not yet use this type.  This would require a great deal more of
refactoring, and will occur as the rest of 'std' is gradually rewritten.